Removed noise (used doubles instead of floats).

+++ camera.cpp
@@ -24,6 +24,6 @@ void Camera::project(float x, float y, Ray &r) const
 {
 	r.origin = pos;
 
-	r.dir = x * u + y * v + w;
+	r.dir = (double)x * u + (double)y * v + w;
 	r.dir.normalize();
 }
