+++ camera.h
@@ -15,14 +15,14 @@ public:
 	void		project(float x, float y, Ray &r) const;
 
 public:
-	Vec3f	pos;
-	Vec3f	target;
-	Vec3f	up;
+	Vec3d	pos;
+	Vec3d	target;
+	Vec3d	up;
 	float	fov;
 
 private:
 	// Define the camera frame
-	Vec3f u, v, w;
+	Vec3d u, v, w;
 };
 
 #endif
