+++ geometry.h
@@ -9,20 +9,20 @@
 class Geometry
 {
 public:
-	virtual bool			hit(const Ray &ray, float &t) const = 0;
-	virtual void			normalAt(const Vec3f &p, Vec3f &n) const = 0;
+	virtual bool			hit(const Ray &ray, double &t) const = 0;
+	virtual void			normalAt(const Vec3d &p, Vec3d &n) const = 0;
 
-	const Vec3f&			position() const	{ return pos; }
+	const Vec3d&			position() const	{ return pos; }
 
 	const MaterialPtr		material() const	{ return mat; }
 	MaterialPtr&			material()			{ return mat; }
 
 protected:
 	Geometry()								{}
-	Geometry(const Vec3f &p) : pos(p)		{}
+	Geometry(const Vec3d &p) : pos(p)		{}
 
 protected:
-	Vec3f			pos;
+	Vec3d			pos;
 	MaterialPtr		mat;
 };
