+++ plane.h
@@ -6,16 +6,16 @@
 class Plane : public Geometry
 {
 public:
-	Plane(const Vec3f &normal = Vec3f(0, 1, 0), float offset = 0, bool _oneSided = true)
+	Plane(const Vec3d &normal = Vec3d(0, 1, 0), float offset = 0, bool _oneSided = true)
 		: n(normal), d(offset), oneSided(_oneSided)
 	{}
 
-	virtual bool			hit(const Ray &ray, float &t) const;
-	virtual void			normalAt(const Vec3f &p, Vec3f &n) const	{ n = this->n; }
+	virtual bool			hit(const Ray &ray, double &t) const;
+	virtual void			normalAt(const Vec3d &p, Vec3d &n) const	{ n = this->n; }
 
 protected:
-	Vec3f		n;
-	float		d;
+	Vec3d		n;
+	double		d;
 	bool		oneSided;
 };
